APPLE-66 Restructure panel sections, create separate metadata section
