+++ assets/js/pluginsidebar/index.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 
 // Components.
 import Icon from '../components/icon';
-import Sidebar from './components/sidebar';
+import Sidebar from './sidebar';
 
 registerPlugin('publish-to-apple-news', {
   icon: <Icon />,
