+++ assets/js/pluginsidebar/panels/maturity-rating.jsx
