+++ assets/js/pluginsidebar/panels/publish-controls.jsx
