+++ assets/js/pluginsidebar/panels/publish-info.jsx
