+++ assets/js/pluginsidebar/panels/pull-quote.jsx
